Math parser updates for size/return
Now keeps track of the max size specified in the parameters. For
example, $0000 + 1 will have a size of 2, meaning it should take up two
individual bytes (01 00)

The return value for math_parse was also changed to be greater than the
max value for a 24 bit integer to prevent conflicts with -1 which is a
valid value.