Massive instruction changes
The instruction struct now handles assembly and storing values. If an
instruction is given a purely mathematical parameter (e.g., AND #$20 +
2), the parameter is sent to the math parser, its size and value are
saved, and the instruction is marked as static which means the value
will not change.

If an instruction is static, it will assemble itself into bytes to be
extracted. If it is not static, (e.g., something like BPL label) it will
not be assembled automatically. Instead, it must be passed a map of
labels that will be forwarded to the math parser to get the value.

As of right now, the non-static instructions are not parsed.